pennypincher/requirements.txt
Django
python-dotenv
django-colorfield